move multi_control_rotation_gates to the synthesis library
