+++ qiskit/circuit/quantumcircuit.py
@@ -35,8 +35,10 @@ from typing import (
     Any,
     DefaultDict,
     Literal,
+    List,
     overload,
 )
+from math import pi
 import numpy as np
 from qiskit._accelerate.circuit import CircuitData
 from qiskit._accelerate.circuit import StandardGate
@@ -4682,6 +4684,199 @@ class QuantumCircuit:
             [],
             copy=False,
         )
+
+    def mcrx(
+        self,
+        theta: ParameterValueType,
+        q_controls: Union[QuantumRegister, List[Qubit]],
+        q_target: Qubit,
+        use_basis_gates: bool = False,
+    ):
+        """
+        Apply Multiple-Controlled X rotation gate
+
+        Args:
+            self (QuantumCircuit): The QuantumCircuit object to apply the mcrx gate on.
+            theta (float): angle theta
+            q_controls (QuantumRegister or list(Qubit)): The list of control qubits
+            q_target (Qubit): The target qubit
+            use_basis_gates (bool): use p, u, cx
+        """
+        from .library.standard_gates.rx import RXGate
+        from qiskit.synthesis.multi_controlled import (
+            _apply_cu,
+            _apply_mcu_graycode,
+            _mcsu2_real_diagonal,
+        )
+
+        control_qubits = self._qbit_argument_conversion(q_controls)
+        target_qubit = self._qbit_argument_conversion(q_target)
+        if len(target_qubit) != 1:
+            raise QiskitError("The mcrz gate needs a single qubit as target.")
+        all_qubits = control_qubits + target_qubit
+        target_qubit = target_qubit[0]
+        self._check_dups(all_qubits)
+
+        n_c = len(control_qubits)
+        if n_c == 1:  # cu
+            _apply_cu(
+                self,
+                theta,
+                -pi / 2,
+                pi / 2,
+                control_qubits[0],
+                target_qubit,
+                use_basis_gates=use_basis_gates,
+            )
+        elif n_c < 4:
+            theta_step = theta * (1 / (2 ** (n_c - 1)))
+            _apply_mcu_graycode(
+                self,
+                theta_step,
+                -pi / 2,
+                pi / 2,
+                control_qubits,
+                target_qubit,
+                use_basis_gates=use_basis_gates,
+            )
+        else:
+            cgate = _mcsu2_real_diagonal(
+                RXGate(theta),
+                num_controls=len(control_qubits),
+                use_basis_gates=use_basis_gates,
+            )
+            self.compose(cgate, control_qubits + [target_qubit], inplace=True)
+
+    def mcry(
+        self,
+        theta: ParameterValueType,
+        q_controls: Union[QuantumRegister, List[Qubit]],
+        q_target: Qubit,
+        q_ancillae: Optional[Union[QuantumRegister, Tuple[QuantumRegister, int]]] = None,
+        mode: Optional[str] = None,
+        use_basis_gates: bool = False,
+    ):
+        """
+        Apply Multiple-Controlled Y rotation gate
+
+        Args:
+            self (QuantumCircuit): The QuantumCircuit object to apply the mcry gate on.
+            theta (float): angle theta
+            q_controls (list(Qubit)): The list of control qubits
+            q_target (Qubit): The target qubit
+            q_ancillae (QuantumRegister or tuple(QuantumRegister, int)): The list of ancillary qubits.
+            mode (string): The implementation mode to use
+            use_basis_gates (bool): use p, u, cx
+
+        """
+        from .library.standard_gates.ry import RYGate
+        from .library.standard_gates.x import MCXGate
+        from qiskit.synthesis.multi_controlled import (
+            _apply_cu,
+            _apply_mcu_graycode,
+            _mcsu2_real_diagonal,
+        )
+
+        control_qubits = self._qbit_argument_conversion(q_controls)
+        target_qubit = self._qbit_argument_conversion(q_target)
+        if len(target_qubit) != 1:
+            raise QiskitError("The mcrz gate needs a single qubit as target.")
+        ancillary_qubits = [] if q_ancillae is None else self._qbit_argument_conversion(q_ancillae)
+        all_qubits = control_qubits + target_qubit + ancillary_qubits
+        target_qubit = target_qubit[0]
+        self._check_dups(all_qubits)
+
+        # auto-select the best mode
+        if mode is None:
+            # if enough ancillary qubits are provided, use the 'v-chain' method
+            additional_vchain = MCXGate.get_num_ancilla_qubits(len(control_qubits), "v-chain")
+            if len(ancillary_qubits) >= additional_vchain:
+                mode = "basic"
+            else:
+                mode = "noancilla"
+
+        if mode == "basic":
+            self.ry(theta / 2, q_target)
+            self.mcx(q_controls, q_target, q_ancillae, mode="v-chain")
+            self.ry(-theta / 2, q_target)
+            self.mcx(q_controls, q_target, q_ancillae, mode="v-chain")
+        elif mode == "noancilla":
+            n_c = len(control_qubits)
+            if n_c == 1:  # cu
+                _apply_cu(
+                    self,
+                    theta,
+                    0,
+                    0,
+                    control_qubits[0],
+                    target_qubit,
+                    use_basis_gates=use_basis_gates,
+                )
+            elif n_c < 4:
+                theta_step = theta * (1 / (2 ** (n_c - 1)))
+                _apply_mcu_graycode(
+                    self,
+                    theta_step,
+                    0,
+                    0,
+                    control_qubits,
+                    target_qubit,
+                    use_basis_gates=use_basis_gates,
+                )
+            else:
+                cgate = _mcsu2_real_diagonal(
+                    RYGate(theta),
+                    num_controls=len(control_qubits),
+                    use_basis_gates=use_basis_gates,
+                )
+                self.compose(cgate, control_qubits + [target_qubit], inplace=True)
+        else:
+            raise QiskitError(f"Unrecognized mode for building MCRY circuit: {mode}.")
+
+    def mcrz(
+        self,
+        lam: ParameterValueType,
+        q_controls: Union[QuantumRegister, List[Qubit]],
+        q_target: Qubit,
+        use_basis_gates: bool = False,
+    ):
+        """
+        Apply Multiple-Controlled Z rotation gate
+
+        Args:
+            self (QuantumCircuit): The QuantumCircuit object to apply the mcrz gate on.
+            lam (float): angle lambda
+            q_controls (list(Qubit)): The list of control qubits
+            q_target (Qubit): The target qubit
+            use_basis_gates (bool): use p, u, cx
+        """
+        from .library.standard_gates.rz import CRZGate, RZGate
+        from qiskit.synthesis.multi_controlled import _mcsu2_real_diagonal
+
+        control_qubits = self._qbit_argument_conversion(q_controls)
+        target_qubit = self._qbit_argument_conversion(q_target)
+        if len(target_qubit) != 1:
+            raise QiskitError("The mcrz gate needs a single qubit as target.")
+        all_qubits = control_qubits + target_qubit
+        target_qubit = target_qubit[0]
+        self._check_dups(all_qubits)
+
+        n_c = len(control_qubits)
+        if n_c == 1:
+            if use_basis_gates:
+                self.u(0, 0, lam / 2, target_qubit)
+                self.cx(control_qubits[0], target_qubit)
+                self.u(0, 0, -lam / 2, target_qubit)
+                self.cx(control_qubits[0], target_qubit)
+            else:
+                self.append(CRZGate(lam), control_qubits + [target_qubit])
+        else:
+            cgate = _mcsu2_real_diagonal(
+                RZGate(lam),
+                num_controls=len(control_qubits),
+                use_basis_gates=use_basis_gates,
+            )
+            self.compose(cgate, control_qubits + [target_qubit], inplace=True)
 
     def r(
         self, theta: ParameterValueType, phi: ParameterValueType, qubit: QubitSpecifier
