+++ qiskit/synthesis/multi_controlled/__init__.py
@@ -22,3 +22,4 @@ from .mcx_synthesis import (
     synth_c3x,
     synth_c4x,
 )
+from .multi_control_rotation_gates import _apply_cu, _apply_mcu_graycode, _mcsu2_real_diagonal
