+++ qiskit/synthesis/multi_controlled/multi_control_rotation_gates.py
@@ -1,6 +1,6 @@
 # This code is part of Qiskit.
 #
-# (C) Copyright IBM 2018, 2019.
+# (C) Copyright IBM 2018, 2024.
 #
 # This code is licensed under the Apache License, Version 2.0. You may
 # obtain a copy of this license in the LICENSE.txt file in the root directory
@@ -10,18 +10,15 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 """
-Multiple-Controlled U3 gate. Not using ancillary qubits.
+Multiple-Controlled U3 gate utilities. Not using ancillary qubits.
 """
 
-from math import pi
 import math
-from typing import Optional, Union, Tuple, List
+from typing import Optional
 import numpy as np
 
-from qiskit.circuit import QuantumCircuit, QuantumRegister, Qubit, Gate
-from qiskit.circuit.library.standard_gates.x import MCXGate
+from qiskit.circuit import QuantumCircuit, Gate
 from qiskit.circuit.library.standard_gates.u3 import _generate_gray_code
-from qiskit.circuit.parameterexpression import ParameterValueType
 from qiskit.exceptions import QiskitError
 
 
@@ -113,7 +110,7 @@ def _mcsu2_real_diagonal(
 
     """
     # pylint: disable=cyclic-import
-    from .x import MCXVChain
+    from qiskit.circuit.library.standard_gates.x import MCXVChain
     from qiskit.circuit.library.standard_gates import RXGate, RYGate, RZGate
     from qiskit.circuit.library.generalized_gates import UnitaryGate
     from qiskit.quantum_info.operators.predicates import is_unitary_matrix
@@ -222,186 +219,3 @@ def _mcsu2_real_diagonal(
         circuit = transpile(circuit, basis_gates=["p", "u", "cx"], qubits_initially_zero=False)
 
     return circuit
-
-
-def mcrx(
-    self,
-    theta: ParameterValueType,
-    q_controls: Union[QuantumRegister, List[Qubit]],
-    q_target: Qubit,
-    use_basis_gates: bool = False,
-):
-    """
-    Apply Multiple-Controlled X rotation gate
-
-    Args:
-        self (QuantumCircuit): The QuantumCircuit object to apply the mcrx gate on.
-        theta (float): angle theta
-        q_controls (QuantumRegister or list(Qubit)): The list of control qubits
-        q_target (Qubit): The target qubit
-        use_basis_gates (bool): use p, u, cx
-    """
-    from .rx import RXGate
-
-    control_qubits = self._qbit_argument_conversion(q_controls)
-    target_qubit = self._qbit_argument_conversion(q_target)
-    if len(target_qubit) != 1:
-        raise QiskitError("The mcrz gate needs a single qubit as target.")
-    all_qubits = control_qubits + target_qubit
-    target_qubit = target_qubit[0]
-    self._check_dups(all_qubits)
-
-    n_c = len(control_qubits)
-    if n_c == 1:  # cu
-        _apply_cu(
-            self,
-            theta,
-            -pi / 2,
-            pi / 2,
-            control_qubits[0],
-            target_qubit,
-            use_basis_gates=use_basis_gates,
-        )
-    elif n_c < 4:
-        theta_step = theta * (1 / (2 ** (n_c - 1)))
-        _apply_mcu_graycode(
-            self,
-            theta_step,
-            -pi / 2,
-            pi / 2,
-            control_qubits,
-            target_qubit,
-            use_basis_gates=use_basis_gates,
-        )
-    else:
-        cgate = _mcsu2_real_diagonal(
-            RXGate(theta),
-            num_controls=len(control_qubits),
-            use_basis_gates=use_basis_gates,
-        )
-        self.compose(cgate, control_qubits + [target_qubit], inplace=True)
-
-
-def mcry(
-    self,
-    theta: ParameterValueType,
-    q_controls: Union[QuantumRegister, List[Qubit]],
-    q_target: Qubit,
-    q_ancillae: Optional[Union[QuantumRegister, Tuple[QuantumRegister, int]]] = None,
-    mode: Optional[str] = None,
-    use_basis_gates: bool = False,
-):
-    """
-    Apply Multiple-Controlled Y rotation gate
-
-    Args:
-        self (QuantumCircuit): The QuantumCircuit object to apply the mcry gate on.
-        theta (float): angle theta
-        q_controls (list(Qubit)): The list of control qubits
-        q_target (Qubit): The target qubit
-        q_ancillae (QuantumRegister or tuple(QuantumRegister, int)): The list of ancillary qubits.
-        mode (string): The implementation mode to use
-        use_basis_gates (bool): use p, u, cx
-
-    """
-    from .ry import RYGate
-
-    control_qubits = self._qbit_argument_conversion(q_controls)
-    target_qubit = self._qbit_argument_conversion(q_target)
-    if len(target_qubit) != 1:
-        raise QiskitError("The mcrz gate needs a single qubit as target.")
-    ancillary_qubits = [] if q_ancillae is None else self._qbit_argument_conversion(q_ancillae)
-    all_qubits = control_qubits + target_qubit + ancillary_qubits
-    target_qubit = target_qubit[0]
-    self._check_dups(all_qubits)
-
-    # auto-select the best mode
-    if mode is None:
-        # if enough ancillary qubits are provided, use the 'v-chain' method
-        additional_vchain = MCXGate.get_num_ancilla_qubits(len(control_qubits), "v-chain")
-        if len(ancillary_qubits) >= additional_vchain:
-            mode = "basic"
-        else:
-            mode = "noancilla"
-
-    if mode == "basic":
-        self.ry(theta / 2, q_target)
-        self.mcx(q_controls, q_target, q_ancillae, mode="v-chain")
-        self.ry(-theta / 2, q_target)
-        self.mcx(q_controls, q_target, q_ancillae, mode="v-chain")
-    elif mode == "noancilla":
-        n_c = len(control_qubits)
-        if n_c == 1:  # cu
-            _apply_cu(
-                self, theta, 0, 0, control_qubits[0], target_qubit, use_basis_gates=use_basis_gates
-            )
-        elif n_c < 4:
-            theta_step = theta * (1 / (2 ** (n_c - 1)))
-            _apply_mcu_graycode(
-                self,
-                theta_step,
-                0,
-                0,
-                control_qubits,
-                target_qubit,
-                use_basis_gates=use_basis_gates,
-            )
-        else:
-            cgate = _mcsu2_real_diagonal(
-                RYGate(theta),
-                num_controls=len(control_qubits),
-                use_basis_gates=use_basis_gates,
-            )
-            self.compose(cgate, control_qubits + [target_qubit], inplace=True)
-    else:
-        raise QiskitError(f"Unrecognized mode for building MCRY circuit: {mode}.")
-
-
-def mcrz(
-    self,
-    lam: ParameterValueType,
-    q_controls: Union[QuantumRegister, List[Qubit]],
-    q_target: Qubit,
-    use_basis_gates: bool = False,
-):
-    """
-    Apply Multiple-Controlled Z rotation gate
-
-    Args:
-        self (QuantumCircuit): The QuantumCircuit object to apply the mcrz gate on.
-        lam (float): angle lambda
-        q_controls (list(Qubit)): The list of control qubits
-        q_target (Qubit): The target qubit
-        use_basis_gates (bool): use p, u, cx
-    """
-    from .rz import CRZGate, RZGate
-
-    control_qubits = self._qbit_argument_conversion(q_controls)
-    target_qubit = self._qbit_argument_conversion(q_target)
-    if len(target_qubit) != 1:
-        raise QiskitError("The mcrz gate needs a single qubit as target.")
-    all_qubits = control_qubits + target_qubit
-    target_qubit = target_qubit[0]
-    self._check_dups(all_qubits)
-
-    n_c = len(control_qubits)
-    if n_c == 1:
-        if use_basis_gates:
-            self.u(0, 0, lam / 2, target_qubit)
-            self.cx(control_qubits[0], target_qubit)
-            self.u(0, 0, -lam / 2, target_qubit)
-            self.cx(control_qubits[0], target_qubit)
-        else:
-            self.append(CRZGate(lam), control_qubits + [target_qubit])
-    else:
-        cgate = _mcsu2_real_diagonal(
-            RZGate(lam),
-            num_controls=len(control_qubits),
-            use_basis_gates=use_basis_gates,
-        )
-        self.compose(cgate, control_qubits + [target_qubit], inplace=True)
-
-
-QuantumCircuit.mcrx = mcrx
-QuantumCircuit.mcry = mcry
-QuantumCircuit.mcrz = mcrz
